Update tests.js
The original index.html that's given to the student contains this UL:
<ul>
<li>
<span><i class="fa fa-trash"></i></span> Eat
</li>
<li>
<span><i class="fa fa-trash"></i></span> Drink
</li>
<li>
<span><i class="fa fa-trash"></i></span> Sleep
</li>
</ul>

Which is 3 <li> items. JEST tests for the function to add a new to-do item, a new <li> to the <ul> and therefore the total number will then be 4. However, currently jest is setup to expect 3.

The last JEST test is for the function to remove a to-do item, to remove a <li> from the parent <ul>, and due to the previous function's additional <li>, it should expect 3, however, currently jest is setup to expect 2. 

To pass the test, right now, students will have to delete an <li> item from the index.html but students are advised to not do so.